requirements.txt
PyYAML
Jinja2
python-dotenv